evals/etl/pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
python-dotenv = "^1.0.1"
langsmith = "^0.1.129"
langchain = "^0.3.1"
pydantic = "^2.9.2"
numpy = "^1.26.4"
langchain-openai = "^0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
